# api-gateway/internal/http/docs/swagger.yaml
definitions:
...
  genproto.ByEmail:
    properties:
      email:
        type: string
    type: object
...
  genproto.MessageSendBCC:
    properties:
      email:
        items:
          type: string
        type: array
    type: object
  genproto.MessageSendCC:
    properties:
      email:
        items:
          type: string
        type: array
    type: object
  genproto.MessageSendTo:
    properties:
      email:
        items:
          type: string
        type: array
    type: object
  genproto.MessageSentRes:
    properties:
      failed_emails:
        items:
          type: string
        type: array
      total_failed:
        type: integer
      total_sent:
        type: integer
    type: object
  genproto.OutboxMessageGetRes:
    properties:
      attachment_ids:
        items:
          type: string
        type: array
      body:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      is_archived:
        type: boolean
      is_draft:
        type: boolean
      is_starred:
        type: boolean
      receivers:
        $ref: '#/definitions/genproto.Receivers'
      sender:
        $ref: '#/definitions/genproto.UserGetRes'
      sent_at:
        type: string
      subject:
        type: string
    type: object
  genproto.OutboxMessageSentBody:
    properties:
      attachment_ids:
        items:
          type: string
        type: array
      body:
        type: string
      receivers:
        $ref: '#/definitions/genproto.Receivers'
      subject:
        type: string
    type: object
  genproto.OutboxMessageSentReq:
    properties:
      body:
        $ref: '#/definitions/genproto.OutboxMessageSentBody'
      sender_id:
        type: string
    type: object
  genproto.OutboxMessagesGetAllRes:
    properties:
      messages:
        items:
          $ref: '#/definitions/genproto.OutboxMessageGetRes'
        type: array
    type: object
...
  genproto.Receivers:
    properties:
      bcc:
        $ref: '#/definitions/genproto.MessageSendBCC'
      cc:
        $ref: '#/definitions/genproto.MessageSendCC'
      to:
        $ref: '#/definitions/genproto.MessageSendTo'
    type: object
...
  genproto.UserGetRes:
    properties:
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      pfp_url:
        type: string
    type: object
  genproto.UserRecoverPasswordReq:
    properties:
      code:
        type: string
      email:
        type: string
      new_password:
        type: string
    type: object
...
  genproto.Void:
    type: object
...
info:
  contact: {}
paths:
...
  /forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a confirmation code to email recovery password
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/genproto.ByEmail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Page not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Forgot passwrod
      tags:
      - 02-Password-recovery
...
  /outbox:
    get:
      consumes:
      - application/json
      description: Retrieves all outbox messages for the authenticated user.
      parameters:
      - description: Search query
        in: query
        name: query
        type: string
      - description: Filter by archived status
        in: query
        name: is_archived
        type: boolean
      - description: Filter by starred status
        in: query
        name: is_starred
        type: boolean
      - description: Filter by sent date (from)
        in: query
        name: sent_from
        type: string
      - description: Filter by sent date (to)
        in: query
        name: sent_to
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of messages per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Outbox messages retrieved successfully
          schema:
            $ref: '#/definitions/genproto.OutboxMessagesGetAllRes'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all outbox messages
      tags:
      - 06-Outbox
    post:
      consumes:
      - application/json
      description: Sends a new message to the specified recipients.
      parameters:
      - description: Message sending request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/genproto.OutboxMessageSentReq'
      produces:
      - application/json
      responses:
        "201":
          description: Message sent successfully
          schema:
            $ref: '#/definitions/genproto.MessageSentRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Send a new message
      tags:
      - 06-Outbox
  /outbox/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently deletes an outbox message.
      parameters:
      - description: Outbox message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Outbox message deleted successfully
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Outbox message not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete outbox message
      tags:
      - 06-Outbox
    get:
      consumes:
      - application/json
      description: Retrieves an outbox message by its ID.
      parameters:
      - description: Outbox message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Outbox message retrieved successfully
          schema:
            $ref: '#/definitions/genproto.OutboxMessageGetRes'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Outbox message not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get outbox message by ID
      tags:
      - 06-Outbox
  /outbox/{id}/archive:
    put:
      consumes:
      - application/json
      description: Archives or unarchives an outbox message.
      parameters:
      - description: Outbox message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Outbox message archived/unarchived successfully
          schema:
            $ref: '#/definitions/genproto.Void'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Outbox message not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Archive outbox message
      tags:
      - 06-Outbox
  /outbox/{id}/star:
    put:
      consumes:
      - application/json
      description: Stars or unstars an outbox message.
      parameters:
      - description: Outbox message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Outbox message starred/unstarred successfully
          schema:
            $ref: '#/definitions/genproto.Void'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Outbox message not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Star outbox message
      tags:
      - 06-Outbox
  /outbox/{id}/trash:
    put:
      consumes:
      - application/json
      description: Moves an outbox message to the trash folder.
      parameters:
      - description: Outbox message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Outbox message moved to trash successfully
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Outbox message not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Move outbox message to trash
      tags:
      - 06-Outbox
...
  /recover-password:
    post:
      consumes:
      - application/json
      description: Verifies the code and updates the password
      parameters:
      - description: Recover Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.UserRecoverPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: Password successfully updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "401":
          description: Incorrect verification code
          schema:
            type: string
        "404":
          description: Verification code expired or email not found
          schema:
            type: string
        "500":
          description: Error updating password
          schema:
            type: string
      summary: Recover password (Use this one after sending verification code)
      tags:
      - 02-Password-recovery
...
  /send-code-again:
    post:
      consumes:
      - application/json
      description: Sends a confirmation code to email recovery password again
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/genproto.ByEmail'
      produces:
      - application/json
      responses:
        "200":
          description: Code sent
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Page not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Sends code again if user didn't recieve the code
      tags:
      - 02-Password-recovery
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...